more dataclass refactor /split out instructions.py
